Add ng-packagr
