--- css/date.component.css
@@ -1,157 +1,0 @@
-.ngx-picker--btn {
-  color: #000;
-  background-color: transparent;
-  border: none;
-  padding: 1em;
-  font-size: 1em; }
-  .ngx-picker--btn__selected {
-    background-color: #eee; }
-  .ngx-picker--btn:hover {
-    transition: .25s ease;
-    background-color: #eee;
-    cursor: pointer; }
-  .ngx-picker--btn:focus {
-    outline-color: transparent; }
-
-.ngx-picker--btn__month:hover {
-  border-top-left-radius: 0.2em; }
-
-.ngx-picker--btn__done:hover {
-  border-bottom-right-radius: 0.2em; }
-
-.ngx-picker--btn__now:hover {
-  border-bottom-left-radius: 0.2em; }
-
-.ngx-picker--btn__next:hover {
-  border-top-right-radius: 0.2em; }
-
-.calendar {
-  position: absolute;
-  background-color: #e6e6e6;
-  max-width: 16em;
-  min-width: 16em;
-  font-size: 16px;
-  font-family: "Arial", sans-serif;
-  border-radius: 0.2em;
-  box-shadow: 0 19px 38px rgba(0, 0, 0, 0.3), 0 15px 12px rgba(0, 0, 0, 0.22); }
-
-.calendar--previous-and-next {
-  position: absolute;
-  right: 0;
-  top: 0;
-  display: flex;
-  justify-content: flex-end; }
-
-.calendar--footer {
-  display: flex;
-  justify-content: space-between;
-  background-color: #e6e6e6;
-  padding: 0;
-  border-radius: 0.2em; }
-
-.calendar--years-select,
-.calendar--months-select {
-  overflow-y: auto;
-  overflow-x: hidden;
-  max-height: 14.85em;
-  background-color: #fff;
-  width: 100%; }
-
-.calendar--year,
-.calendar--month {
-  display: block;
-  padding: .5em 1em;
-  width: 100%;
-  overflow: hidden;
-  color: #000; }
-  .calendar--year__selected,
-  .calendar--month__selected {
-    color: #000;
-    font-weight: bold;
-    background-color: #eee; }
-  .calendar--year:hover,
-  .calendar--month:hover {
-    cursor: pointer;
-    background-color: #eee;
-    font-weight: 400;
-    color: #000; }
-
-.calendar--days-of-week {
-  display: flex;
-  justify-content: space-around;
-  background-color: #eee; }
-
-.calendar--day-of-week {
-  position: relative;
-  padding: .5em;
-  text-align: center;
-  text-transform: uppercase;
-  color: #000;
-  font-size: .9em; }
-
-.calendar--days-select {
-  display: flex;
-  flex-wrap: wrap;
-  flex-direction: row;
-  justify-content: space-between;
-  background-color: #fff; }
-
-.calendar--day {
-  text-align: center;
-  width: 2.25em;
-  padding: .5em 0;
-  font-weight: 500;
-  color: #000; }
-  .calendar--day__muted {
-    color: #999; }
-  .calendar--day__selected {
-    background-color: #eee;
-    color: #000; }
-  .calendar--day:hover {
-    cursor: pointer;
-    color: #000;
-    background-color: #eee; }
-
-.ngx-picker {
-  display: flex;
-  position: relative; }
-  .ngx-picker .calendar {
-    position: absolute;
-    left: 0;
-    top: 2.5em;
-    z-index: 3; }
-
-.time-picker {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex-wrap: wrap;
-  max-width: 16em;
-  min-width: 16em;
-  background-color: #e6e6e6; }
-
-.time--values {
-  display: flex;
-  justify-content: space-between;
-  flex-direction: row;
-  flex-wrap: wrap;
-  max-width: 16em;
-  min-width: 16em;
-  background-color: #eee; }
-
-.time--value {
-  width: 2.66em;
-  font-size: 1em;
-  padding: .5em 0;
-  text-align: center;
-  color: #000; }
-  .time--value__selected {
-    color: #000;
-    background-color: #e6e6e6; }
-  .time--value:hover {
-    transition: .25s ease;
-    cursor: pointer;
-    background-color: #e6e6e6; }
-
-[hidden] {
-  display: none !important; }
