+++ demo/package.json
@@ -20,7 +20,7 @@
     "@angular/platform-browser-dynamic": "^8.0.0",
     "@angular/router": "^8.0.0",
     "core-js": "^2.4.1",
-    "ngx-datetime-picker": "file:../dist/ngx-datetime-picker-2.0.0.tgz",
+    "ngx-datetime-picker": "^2.1.3",
     "rxjs": "^6.5.2",
     "rxjs-compat": "^6.0.0-rc.0",
     "tslib": "^1.9.0",
@@ -42,7 +42,7 @@
     "karma-remap-istanbul": "^0.2.1",
     "protractor": "~4.0.13",
     "ts-node": "1.2.1",
-    "tslint": "^4.3.0",
+    "tslint": "^5.20.0",
     "typescript": "~3.4.5"
   }
 }
