+++ e2e/tsconfig.json
@@ -10,7 +10,7 @@
     "sourceMap": true,
     "target": "es5",
     "typeRoots": [
-      "../node_modules/@types"
+      "../ngx-datetimepicker/node_modules/@types"
     ]
   }
 }
