--- ngx-datetimepicker/package.json
@@ -1,45 +1,0 @@
-{
-    "name": "ngx-datetimepicker",
-    "version": "0.0.0",
-    "license": "MIT",
-    "angular-cli": {},
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "test": "ng test",
-        "pree2e": "webdriver-manager update --standalone false --gecko false",
-        "e2e": "protractor"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/common": "^4.0.1",
-        "@angular/compiler": "^4.0.1",
-        "@angular/core": "^4.0.1",
-        "@angular/forms": "^4.0.1",
-        "@angular/http": "^4.0.1",
-        "@angular/platform-browser": "^4.0.1",
-        "@angular/platform-browser-dynamic": "^4.0.1",
-        "@angular/router": "^4.0.1",
-        "core-js": "^2.4.1",
-        "rxjs": "^5.1.0",
-        "zone.js": "^0.8.4"
-    },
-    "devDependencies": {
-        "@angular/cli": "^1.0.0",
-        "@angular/compiler-cli": "^4.0.1",
-        "@types/jasmine": "2.5.38",
-        "@types/node": "^6.0.42",
-        "codelyzer": "~2.0.0-beta.1",
-        "jasmine-core": "2.5.2",
-        "jasmine-spec-reporter": "2.5.0",
-        "karma": "1.2.0",
-        "karma-chrome-launcher": "^2.0.0",
-        "karma-cli": "^1.0.1",
-        "karma-jasmine": "^1.0.2",
-        "karma-remap-istanbul": "^0.2.1",
-        "protractor": "~4.0.13",
-        "ts-node": "1.2.1",
-        "tslint": "^4.3.0",
-        "typescript": "~2.2.2"
-    }
-}
