+++ src/app/models/styleObject.model.ts
