+++ src/app/time.component/time.component.ts
