+++ src/app/timePicker.component/timePicker.component.ts
