+++ src/scss/ngx-datetime-picker.css
