+++ src/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "baseUrl": "",
+    "baseUrl": "tsconfig",
     "declaration": true,
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
@@ -12,7 +12,7 @@
     "target": "es5",
     "typeRoots": [
       "../node_modules/@types",
-      "./typings/isMobile.d.ts"
+        "./typings/isMobile.d.ts"
     ]
   }
 }
